Clean up

+++ vetclinic-server/src/main/kotlin/personal/ciai/vetclinic/service/AdministratorService.kt
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import personal.ciai.vetclinic.config.ConfigurationProperties
 import personal.ciai.vetclinic.dto.AdministratorDTO
+import personal.ciai.vetclinic.exception.AccessForbiddenException
 import personal.ciai.vetclinic.exception.ConflictException
 import personal.ciai.vetclinic.exception.NotFoundException
 import personal.ciai.vetclinic.model.Administrator
@@ -37,6 +38,8 @@ class AdministratorService(
     }
 
     fun delete(id: Int) {
+        if (id==1)
+            throw AccessForbiddenException("This administrator cannot be deleted")
         val admin: Administrator = getAdministratorEntity(id)
 
         adminRepository.delete(admin)
